Rename json_processor and main modules
