+++ docs/src/changelog/index.rst
@@ -4,10 +4,7 @@
 Changelog
 =========
 
-Changes ever since 0.2.0 (because I actually changed so much this time it
-*has* to be addressed)
-
 .. toctree::
    :maxdepth: 2
 
-   0.2.rst
+   0.1.rst
