+++ docs/src/index.rst
@@ -1,8 +1,10 @@
 .. _home:
 
 ======================================
-The Perspective Library documentation!
+The Perspective Journal documentation!
 ======================================
+
+.. TODO: UPDATE THIS calling this a library just drives people off
 
 Welcome!  This is a pure Python library providing backup utility for the
 iOS journaling app `Perspective`_, as well as backing up and processing
