+++ docs/src/psp/ext/index.rst
@@ -1,17 +1,19 @@
-.. _exts:
+.. _ext:
 
-=============================================================================
-:mod:`extensions` --- Extensions that may or may not be so useful to everyone
-=============================================================================
+==========================================================================
+:mod:`psp.ext` --- Extensions that may or may not be so useful to everyone
+==========================================================================
 
 This module stores various additional functionalities beyond the basic needs
 of backing up Perspective entries.  Since not everyone wants all of these
 functionalities overloaded by default, these are broken down into optional
 plugins so you only have to pick what you want!
 
+.. singular ext because i don't know
+
 See the :ref:`Extensions tutorial <exttut>` to get started.
 
-.. module:: psp.extensions
+.. module:: psp.ext
    :synopsis: Miscellaneous extensions
 
 .. toctree::
