+++ docs/src/psp/util.rst
@@ -1,8 +1,11 @@
 .. _util:
 
-=======================================
-:mod:`util` --- Miscellaneous functions
-=======================================
+===========================================
+:mod:`psp.util` --- Miscellaneous functions
+===========================================
+
+.. module:: psp.util
+   :synopsis: Miscellaneous functions
 
 Functions that I either don't know where to put or are
 for debugging purposes (like `cmpentries` and `cmppanels`)
