+++ docs/src/tutorials/cookbook.rst
@@ -47,15 +47,15 @@ archive like this:
    }
 
 By default, the
-:meth:`JSONLoader.load() <psp.processors.json_processor.JSONLoader.load>`
+:meth:`JSONLoader.load() <psp.serializers.json.JSONLoader.load>`
 method only parses top-level attributes like *tz* and *paths* and
 ignores the rest.  This is to give users freedom to put in whatever
 they find useful (description, checksums, etc.) without having to
 filter them out.
 
 That is to say, there is no canonical key for *description* that is
-reinforced by :class:`~psp.processors.json_processor.JSONLoader` or
-:class:`~psp.processors.json_processor.JSONDumper`, although
+reinforced by :class:`~psp.serializers.json.JSONLoader` or
+:class:`~psp.serializers.json.JSONDumper`, although
 *desc* is encouraged and was used in my past code!  As an example,
 the ``get_description()`` function in ``config.py`` generated by
 ``psp init`` tries to query this special attribute::
@@ -72,7 +72,7 @@ the ``get_description()`` function in ``config.py`` generated by
 
 To dump to a file
 
-   >>> from psp.processors import JSONLoader, JSONDumper
+   >>> from psp.serializers import JSONLoader, JSONDumper
    >>> import io
    >>> # Create a sample panel
    >>> (p,) = JSONLoader().load_data(
@@ -139,7 +139,7 @@ Main program: speed?
 ^^^^^^^^^^^^^^^^^^^^
 
 Quick check by skipping the
-:meth:`JSONLoader.load() <psp.processors.json_processor.JSONLoader.load>`
+:meth:`JSONLoader.load() <psp.serializers.json.JSONLoader.load>`
 interface::
 
     def load_singles(path, encoding, date):
