+++ src/psp/__init__.py
@@ -1,5 +1,5 @@
 """The Perspective library."""
-__version__ = '0.2.0a1'
+__version__ = '0.1.0a1'
 __version_info__ = (0, 2, 0)
 
 # Let's not import anything here... a lot of things are
