+++ src/psp/_templates/mystuff.py.txt
@@ -1,5 +1,5 @@
 """my stuff!"""
-from psp.processors import JSONLoader, JSONDumper
+from psp.serializers import JSONLoader, JSONDumper
 from psp.stringify import PanelFormatter
 import psp.timeutil
