--- src/psp/processors/__init__.py
@@ -1,7 +1,0 @@
-"""Processors for backup files."""
-
-__all__ = []
-
-from . import json_processor
-from .json_processor import JSONLoader, JSONDumper, load_json, dump_json
-__all__.extend(['JSONLoader', 'JSONDumper', 'load_json', 'dump_json'])
