+++ src/psp/serializers/__init__.py
@@ -1,0 +1,10 @@
+"""Serialization functions (formerly processors) for backup files."""
+
+__all__ = []
+
+from . import json
+from .json import *
+from . import text
+from .text import *
+__all__.extend(json.__all__)
+__all__.extend(text.__all__)
