+++ tests/test_extensions/test_captions.py
@@ -6,9 +6,10 @@ import shutil
 import tempfile
 import unittest
 
-from psp.extensions import captions
+from psp.ext import captions
 from ..test_processors.test_json_processor import (
-    make_backup, add_entries, write_json, write_file)
+    make_backup, add_entries, write_json, write_file
+)
 
 
 def setUpModule():
